+++ src/mastodon_api/accounts/views.rs
@@ -67,6 +67,7 @@ use crate::mastodon_api::{
 };
 use crate::models::{
     posts::queries::get_posts_by_author,
+    profiles::helpers::find_aliases,
     profiles::queries::{
         get_profile_by_acct,
         get_profile_by_id,
@@ -797,6 +798,28 @@ async fn get_account_subscribers(
     Ok(HttpResponse::Ok().json(subscriptions))
 }
 
+#[get("/{account_id}/aliases")]
+async fn get_account_aliases(
+    connection_info: ConnectionInfo,
+    config: web::Data<Config>,
+    db_pool: web::Data<DbPool>,
+    account_id: web::Path<Uuid>,
+) -> Result<HttpResponse, MastodonError> {
+    let db_client = &**get_database_client(&db_pool).await?;
+    let profile = get_profile_by_id(db_client, &account_id).await?;
+    let aliases = find_aliases(db_client, &profile).await?;
+    let base_url = get_request_base_url(connection_info);
+    let instance_url = config.instance_url();
+    let accounts: Vec<Account> = aliases.into_iter()
+        .map(|profile| Account::from_profile(
+            &base_url,
+            &instance_url,
+            profile,
+        ))
+        .collect();
+    Ok(HttpResponse::Ok().json(accounts))
+}
+
 pub fn account_api_scope() -> Scope {
     web::scope("/api/v1/accounts")
         // Routes without account ID
@@ -819,4 +842,5 @@ pub fn account_api_scope() -> Scope {
         .service(get_account_followers)
         .service(get_account_following)
         .service(get_account_subscribers)
+        .service(get_account_aliases)
 }